+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [1.0.12] - 2022-10-09
 ### Added
 - Warmup-2 stringMatch
+- Warmup-2 stringX
 
 ## [1.0.11] - 2022-10-08
 ### Added